requirements.txt
python-telegram-bot[job-queue]>=21.0
sqlalchemy>=2.0
alembic>=1.10
python-dotenv>=1.0
requests>=2.31
psycopg2-binary>=2.9.9